Assert account existence when adding to cart
Was already in place for `add_contra`, but not for the entry itself.

The assertion was done already earlier in the call stack (in
`$cart->checkout`), but wouldn't be called if someone would manually
call `$entry->account($new)`. Not that anyone would/should do that, but
better safe than sorry.

There's a small but unnoticeable slowdown because the assertion is now
done multiple times, and each time the entire accounts file is read and
parsed.